Feat: 홈 화면 대시보드 전환 및 라우팅 구조 개편 #199
